requirements.txt
flet>=0.21.0
pexpect>=4.8.0
pillow>=10.0.0 